refactor: all folder structure and arch definitions setup
