+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM python:3.12-slim
+FROM python:3.12
 
 # Defina a variável de ambiente para não gerar bytecode compilado
 ENV PYTHONDONTWRITEBYTECODE=1
