+++ pyproject.toml
@@ -20,6 +20,7 @@ pytest = "^8.3.3"
 python-dotenv = "^1.0.1"
 coverage = "^7.6.4"
 tqdm = "^4.66.5"
+pydantic = "^2.10.3"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.2"
