+++ src/infrastructure/services/send_metrics_service.py
@@ -1,6 +1,4 @@
 import os
-from datetime import datetime
-from dotenv import load_dotenv
 from typing import Tuple, Optional
 from datadog_api_client import ApiClient, Configuration
 from datadog_api_client.v2.api.metrics_api import MetricsApi
@@ -8,17 +6,13 @@ from datadog_api_client.v2.model.metric_intake_type import MetricIntakeType
 from datadog_api_client.v2.model.metric_payload import MetricPayload
 from datadog_api_client.v2.model.metric_point import MetricPoint
 from datadog_api_client.v2.model.metric_series import MetricSeries
-from src.application.ports.metrics_interface import MetricsPort
-from src.application.use_cases.get_secrets import GetSecretValueUseCase
-from src.application.utils.logger_module import logger, log_extra_info, LogStatus
-
-load_dotenv()
+from src.infrastructure.utils.logger_module import logger, log_extra_info, LogStatus
+from src.infrastructure.utils.logger_module import logger, log_extra_info, LogStatus
 
 
-class DataDogAPIAdapter(MetricsPort):
-    def __init__(self, metric_name: str):
-        self.metric_name = metric_name
-        self.secret_manager = GetSecretValueUseCase()
+class Datadog:
+    def __init__(self):
+        self._metric_name = "pypi"
         self.configuration = Configuration()
         self.configuration.api_key["apiKeyAuth"] = os.getenv("DATADOG_API_KEY")
         self.configuration.host = os.getenv("DATADOG_HOST")
@@ -26,17 +20,11 @@ class DataDogAPIAdapter(MetricsPort):
         self.configuration.enable_retry = True
         self.configuration.max_retries = 3
 
-    def _get_dd_api_key(self):
-        return self.secret_manager.get(secret_id="datadog-pypi-package-stats")
-
-    def _get_dd_host(self):
-        return self.secret_manager.get(secret_id="datadog-host")
-
     def increment(self, tags: list, value: int, timestamp) -> Tuple[str, Optional[str]]:
         body = MetricPayload(
             series=[
                 MetricSeries(
-                    metric=self.metric_name,
+                    metric=self._metric_name,
                     type=MetricIntakeType.COUNT,
                     points=[
                         MetricPoint(
@@ -75,3 +63,18 @@ class DataDogAPIAdapter(MetricsPort):
             extra=log_extra_info(status=LogStatus.OK),
         )
         return response.to_str(), None
+
+
+class SendMetricsService:
+    def __init__(self):
+        self.metrics_repository = Datadog()
+
+    def send(
+        self,
+        tags: list,
+        value: int,
+        timestamp: float,
+    ) -> Tuple[str, Optional[str]]:
+        return self.metrics_repository.increment(
+            tags=tags, value=value, timestamp=timestamp
+        )
